refactor(onboarding): extract resolveFrameworkIds to shared helper
Deduplicates the resolveFrameworkIds logic that was copied in both
complete-onboarding.ts and initialize-organization.ts. Now lives in
actions/organization/lib/resolve-framework-ids.ts alongside
initialize-organization.ts.
